Add 'store' module + updates in model
