+++ lib/safira/store/buy.ex
@@ -1,12 +1,12 @@
-defmodule Safira.Contest.Buy do
+defmodule Safira.Store.Buy do
   use Ecto.Schema
   import Ecto.Changeset
-  alias Safira.Contest.Redeemable
+  alias Safira.Store.Redeemable
   alias Safira.Accounts.Attendee
 
   schema "buys" do
-    belongs_to :attendee, Attendee, foreign_key: :attendee_id, type: :binary_id
-    belongs_to :redeemable, Redeemable, foreign_key: :redeemable_id, type: :binary_id
+    belongs_to(:attendee, Attendee, foreign_key: :attendee_id, type: :binary_id)
+    belongs_to(:redeemable, Redeemable, foreign_key: :redeemable_id, type: :binary_id)
 
     timestamps()
   end
@@ -16,5 +16,4 @@ defmodule Safira.Contest.Buy do
     |> cast(attrs, [:attendee_id, :redeemable_id])
     |> validate_required([:attendee_id, :redeemable_id])
   end
-
 end
